# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MYE030 Data Integration & Visualization API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/conferences": {
      "get": {
        "tags": [
          "conferences"
        ],
        "summary": "List Conferences",
        "description": "Return a paginated list of conferences, optionally filtered.\n\n``search_text`` is a substring match against both title and acronym.\n``primary_for`` is a substring match. ``ranked_only`` hides rows\nwithout an iCore rank; ``has_acronym`` hides rows without an\nacronym; ``has_for`` hides rows without a Field of Research code.\nSortable columns: ``title``, ``acronym``, ``rank_value``,\n``primary_for``.",
        "operationId": "list_conferences_api_conferences_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "search_text",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search Text"
            }
          },
          {
            "name": "rank_value",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rank Value"
            }
          },
          {
            "name": "primary_for",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Primary For"
            }
          },
          {
            "name": "ranked_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Ranked Only"
            }
          },
          {
            "name": "has_acronym",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Has Acronym"
            }
          },
          {
            "name": "has_for",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Has For"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order By"
            }
          },
          {
            "name": "order_dir",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(asc|desc)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ConferenceSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conferences/{conference_id}": {
      "get": {
        "tags": [
          "conferences"
        ],
        "summary": "Get Conference Profile",
        "description": "Return the aggregated profile of one conference, optionally year-bounded.",
        "operationId": "get_conference_profile_api_conferences__conference_id__get",
        "parameters": [
          {
            "name": "conference_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Conference Id"
            }
          },
          {
            "name": "start_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Year"
            }
          },
          {
            "name": "end_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConferenceProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conferences/{conference_id}/articles/{article_id}": {
      "get": {
        "tags": [
          "conferences"
        ],
        "summary": "Get Conference Article Details",
        "description": "Return full details of one conference paper including its authors.",
        "operationId": "get_conference_article_details_api_conferences__conference_id__articles__article_id__get",
        "parameters": [
          {
            "name": "conference_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Conference Id"
            }
          },
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conferences/{conference_id}/yearly-statistics": {
      "get": {
        "tags": [
          "conferences"
        ],
        "summary": "Get Conference Yearly Statistics",
        "description": "Return the per-year publication statistics that drive the line chart.\n\nTrims a trailing partial year by default.",
        "operationId": "get_conference_yearly_statistics_api_conferences__conference_id__yearly_statistics_get",
        "parameters": [
          {
            "name": "conference_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Conference Id"
            }
          },
          {
            "name": "start_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Year"
            }
          },
          {
            "name": "end_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Year"
            }
          },
          {
            "name": "include_partial_last_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Partial Last Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConferenceYearlyStatistic"
                  },
                  "title": "Response Get Conference Yearly Statistics Api Conferences  Conference Id  Yearly Statistics Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conferences/{conference_id}/articles": {
      "get": {
        "tags": [
          "conferences"
        ],
        "summary": "List Conference Articles",
        "description": "Return a paginated list of articles presented at the conference.\n\nSortable columns: ``title``, ``year``, ``pages``.",
        "operationId": "list_conference_articles_api_conferences__conference_id__articles_get",
        "parameters": [
          {
            "name": "conference_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Conference Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "start_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Year"
            }
          },
          {
            "name": "end_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Year"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order By"
            }
          },
          {
            "name": "order_dir",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(asc|desc)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ConferenceArticle_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/years/{year}/conferences": {
      "get": {
        "tags": [
          "years"
        ],
        "summary": "List Year Conferences",
        "description": "Return conferences that held papers in the requested year.\n\nSortable columns: ``title``, ``acronym``, ``rank_value``,\n``articles_in_year``.",
        "operationId": "list_year_conferences_api_years__year__conferences_get",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2100,
              "minimum": 1900,
              "title": "Year"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order By"
            }
          },
          {
            "name": "order_dir",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(asc|desc)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_YearConferenceEntry_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConferenceProfile": {
        "properties": {
          "conference_id": {
            "type": "integer",
            "title": "Conference Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "acronym": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acronym"
          },
          "rank_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank Value"
          },
          "primary_for": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary For"
          },
          "primary_for_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary For Description"
          },
          "total_articles": {
            "type": "integer",
            "title": "Total Articles"
          },
          "earliest_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earliest Year"
          },
          "latest_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Year"
          },
          "distinct_authors_total": {
            "type": "integer",
            "title": "Distinct Authors Total"
          },
          "average_articles_per_year": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Articles Per Year"
          },
          "average_authors_per_article_overall": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Authors Per Article Overall"
          }
        },
        "type": "object",
        "required": [
          "conference_id",
          "title",
          "total_articles",
          "distinct_authors_total"
        ],
        "title": "ConferenceProfile",
        "description": "Full profile combining iCore ranking columns with aggregate stats."
      },
      "ConferenceSummary": {
        "properties": {
          "conference_id": {
            "type": "integer",
            "title": "Conference Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "acronym": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acronym"
          },
          "rank_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank Value"
          },
          "primary_for": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary For"
          },
          "primary_for_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary For Description"
          }
        },
        "type": "object",
        "required": [
          "conference_id",
          "title"
        ],
        "title": "ConferenceSummary",
        "description": "Lightweight projection used by list endpoints."
      },
      "ConferenceYearlyStatistic": {
        "properties": {
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "articles_count": {
            "type": "integer",
            "title": "Articles Count"
          },
          "distinct_authors": {
            "type": "integer",
            "title": "Distinct Authors"
          },
          "total_authors": {
            "type": "integer",
            "title": "Total Authors"
          },
          "average_authors_per_article": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Authors Per Article"
          }
        },
        "type": "object",
        "required": [
          "year",
          "articles_count",
          "distinct_authors",
          "total_authors"
        ],
        "title": "ConferenceYearlyStatistic",
        "description": "One bucket of the conference's per-year aggregate statistics."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_ConferenceArticle_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConferenceArticle"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Items"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total_items"
        ],
        "title": "PaginatedResponse[ConferenceArticle]"
      },
      "PaginatedResponse_ConferenceSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConferenceSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Items"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total_items"
        ],
        "title": "PaginatedResponse[ConferenceSummary]"
      },
...
      "PaginatedResponse_YearConferenceEntry_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/YearConferenceEntry"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Items"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total_items"
        ],
        "title": "PaginatedResponse[YearConferenceEntry]"
      },
...
      "PaperAuthor": {
        "properties": {
          "author_id": {
            "type": "integer",
            "title": "Author Id"
          },
          "author_name": {
            "type": "string",
            "title": "Author Name"
          }
        },
        "type": "object",
        "required": [
          "author_id",
          "author_name"
        ],
        "title": "PaperAuthor",
        "description": "Author entry inside a paper detail response."
      },
      "PaperDetails": {
        "properties": {
          "article_id": {
            "type": "integer",
            "title": "Article Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "pages": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "venue_type": {
            "type": "string",
            "title": "Venue Type"
          },
          "venue_id": {
            "type": "integer",
            "title": "Venue Id"
          },
          "venue_title": {
            "type": "string",
            "title": "Venue Title"
          },
          "authors": {
            "items": {
              "$ref": "#/components/schemas/PaperAuthor"
            },
            "type": "array",
            "title": "Authors"
          }
        },
        "type": "object",
        "required": [
          "article_id",
          "title",
          "venue_type",
          "venue_id",
          "venue_title",
          "authors"
        ],
        "title": "PaperDetails",
        "description": "Full information about a single article including its authors."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}